+++ Source/SwiftLintFramework/Models/MasterRuleList.swift
@@ -71,6 +71,7 @@ public let masterRuleList = RuleList(rules: [
     LiteralExpressionEndIdentationRule.self,
     LowerACLThanParentRule.self,
     MarkRule.self,
+    ModifiersOrderRule.self,
     MultilineArgumentsRule.self,
     MultilineParametersRule.self,
     MultipleClosuresWithTrailingClosureRule.self,