+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 #### Enhancements
 
+* Added `modifierOrderRule` to enforce the order of modifiers.  
++  [Jose Cheyo Jimenez](https://github.com/masters3d)
++  [Daniel Metzing](https://github.com/dirtydanee)
++  [#1472](https://github.com/realm/SwiftLint/issues/1472)
+
 * Validate implicit `subscript` getter in `implicit_getter` rule when using
   Swift 4.1.  
   [Marcelo Fabri](https://github.com/marcelofabri)